Lora stack refactor

+++ lib/initialisation.py
@@ -1,17 +1,15 @@
 from RtcDS1307 import clock
 import GpsSIM28
-from network import LoRa
-import socket
-import ubinascii
-from Configuration import config
-from PM_thread import pm_thread
+from PM_read import pm_thread
 from SensorLogger import SensorLogger
+from loggingpycom import INFO, WARNING, CRITICAL, DEBUG, ERROR
+from Configuration import config
 import _thread
 import strings as s
 import os
 
 
-def initialize_time(rtc, logger):
+def initialize_time(rtc, gps_on, logger):
     no_time = False
     update_time_later = True
     try:
@@ -21,7 +19,7 @@ def initialize_time(rtc, logger):
         try:
             if rtc.now()[0] < 2019 or rtc.now()[0] >= 2100:
                 # Get time and calibrate RTC module via GPS
-                if config.get_config("GPS") == "ON":
+                if gps_on:
                     GpsSIM28.get_time(rtc, True, logger)
                     update_time_later = False
                 # Calibrate RTC module via WiFi Configurations and then reboot
@@ -36,7 +34,7 @@ def initialize_time(rtc, logger):
         logger.exception("Failed to get time from RTC module")
         try:
             # Get time via GPS
-            if config.get_config("GPS") == "ON":
+            if gps_on:
                 GpsSIM28.get_time(rtc, True, logger)
                 update_time_later = False
             # No way of getting time
@@ -78,50 +76,37 @@ def initialize_pm_sensor(sensor_name, pins, serial_id, status_logger):
 
 def initialize_file_system():
     """
-    Create directories for logging, processing, and sending data if they do not exist.
+    Create directories for logging, processing, archive, and sending data if they do not exist.
     """
+    # Create directories in /sd/
     for directory in s.filesystem_dirs:
         if directory not in os.listdir(s.root_path):
             os.mkdir(s.root_path + directory)
+
+    # Create Averages directory in /sd/Archive/ directory
+    if s.archive_averages not in os.listdir(s.root_path + s.archive):
+        os.mkdir(s.archive_path + s.archive_averages)
 
 
 def remove_residual_files():
     """
     Removes residual files from the last boot in the current and processing dirs
     """
-    for path in [s.current_path, s.processing_path, s.lora_path]:
+    for path in [s.current_path, s.processing_path]:
         for file in os.listdir(path[:-1]):  # Strip '/' from the end of path
-            os.remove(path + file)
+            if file != s.lora_file_name:
+                os.remove(path + file)
 
 
-def initialize_lorawan():
-    # default region is Europe
-    region = LoRa.EU868
-
-    # set region according to configuration
-    if config.get_config("region") == "Asia":
-        region = LoRa.AS923
-    elif config.get_config("region") == "Australia":
-        region = LoRa.AU915
-    elif config.get_config("region") == "United States":
-        region = LoRa.US915
-
-    lora = LoRa(mode=LoRa.LORAWAN, region=region, adr=True)
-
-    # create an OTAA authentication parameters
-    app_eui = ubinascii.unhexlify(config.get_config("application_eui"))
-    app_key = ubinascii.unhexlify(config.get_config("app_key"))
-
-    # join a network using OTAA (Over the Air Activation)
-    lora.join(activation=LoRa.OTAA, auth=(app_eui, app_key), timeout=0)
-
-    # create a LoRa socket
-    lora_socket = socket.socket(socket.AF_LORA, socket.SOCK_RAW)
-
-    # request acknowledgment of data sent
-    lora_socket.setsockopt(socket.SOL_LORA, socket.SO_CONFIRMED, True)
-
-    # sets timeout for sending data
-    lora_socket.settimeout(int(config.get_config("lora_timeout")) * 1000)
-
-    return lora, lora_socket
+def get_logging_level():
+    logging_lvl = config.get_config("logging_lvl")
+    if logging_lvl == "Critical":
+        return CRITICAL
+    elif logging_lvl == "Error":
+        return ERROR
+    elif logging_lvl == "Warning":
+        return WARNING
+    elif logging_lvl == "Info":
+        return INFO
+    elif logging_lvl == "Debug":
+        return DEBUG
